Added navbar
